Revert "[k8s] Replace config map replace to list + create or update (#1521)"
This reverts commit d34ff4b51cbef2fede42c070c838ad1d818eaee8.

The commit does not compile and should not have been checked in.
